Bayes

+++ workflows/scaling_factor/visualize.py
@@ -104,7 +104,7 @@ def sf_plot(summary, conf_matrix, classes, test_classes, training_method, save_d
             classes_y=text_test_classes,
             axes=acc_axes[i],
             title="Tested using {} - Accuracy = {:.2f}%".format(
-                method_titles[i], float(100*conf_matrix[i].diagonal().mean())),
+                method_titles[i], 100*conf_matrix[i].diagonal().mean()),
             cbar=False,
             cmap="Greys",
         )
